+++ src/app.js
@@ -1,42 +1,42 @@
 const express = require("express");
+const path = require("path"); 
+const methodOverride = require('method-override');
+const session = require('express-session');
 const app = express();
-const path = require("path");
+
+/*****************ROUTERS************************/
 const mainRouter = require("./routes/mainRouter")
 const productRouter = require('./routes/productRouter');
 const userRouter = require('./routes/userRouter');
-const methodOverride = require('method-override');
-const { rmSync } = require("fs");
-const session = require('express-session');
-const cookieParser = require('cookie-parser');
-const validateRegister = require("./middlewares/validateRegister");
-
-
-app.use(methodOverride('_method'));
-
-app.set('view engine','ejs')
-
-app.set('views', path.join(__dirname, '/views'))
-
-app.use(express.static('public'));
-
-app.use("/public", express.static("public"));
-
+/***************ROUTING*******************/
 app.use('/',mainRouter)
-
 app.use('/product', productRouter);
-
 app.use('/user', userRouter);
-
+/*****************SESSION************************/
 app.use(session( {
     secret: "Nuestro mensaje secreto",
     resave: false,
     saveUninitialized: false,
     
 }));
+/**************EDIT DELETE*****************/
+app.use(methodOverride('_method'));
 
+/****************VIEWS*************************/
+app.set('view engine','ejs')
+
+app.set('views', path.join(__dirname, '/views'))
+
+/*****************STYLES***************************/
+app.use(express.static('public'));
+app.use("/public", express.static("public"));
+
+/************FORMS DATA******************/
 app.use(express.urlencoded({ extended: true}));
 app.use(express.json());
 
+
+/****************404**********************/
 app.use((req, res, next) => {
     res.status(404).render('not-found');
 })
@@ -46,6 +46,7 @@ app.listen(3000,(req,res) =>{
 })
 
 module.exports = app;
+
