+++ src/middlewares/authMiddleware.js
@@ -1,8 +1,8 @@
-/*function authMiddleware (req, res, next){
+function authMiddleware (req, res, next){
     if(!req.session.userLogged) {
         return res.redirect ('user/login')
     }
     next();
 }
 
-module.exports = authMiddleware;*/
+module.exports = authMiddleware;
