+++ src/middlewares/guestMiddleware.js
@@ -1,4 +1,4 @@
-/*function guestMiddleware(req, res, next){
+function guestMiddleware(req, res, next){
     if (req.session.userLogged){
         return res.redirect('user/profile');
 
@@ -6,4 +6,4 @@
     next();
 };
 
-module.exports = guestMiddleware; */
+module.exports = guestMiddleware;
