+++ src/middlewares/validateUserEdit.js
@@ -26,17 +26,6 @@ module.exports = [
         .notEmpty() .withMessage('Completa con tu Email') .bail()
         .isEmail() .withMessage('Debes usar un formato de email valido, EJ: usuario@hotmail.com'),
         
-        body('password')
-        .notEmpty() .withMessage('Tienes que ingresar tu contraseña') 
-        .isLength({ min: 8}) .withMessage('(*) Debe contener un mínimo de 8 carateres'),
         
-        body('repassword')
-        .custom((value, { req }) => {
-            if (value !== req.body.password) {
-            throw new Error('(*) Las contraseñas no coinciden');
-            }
-        
-            return true;
-        })
-        
+
 ];
