+++ src/views/partials/header.ejs
@@ -9,9 +9,10 @@
                <!-- <li class="menu"><a href="#"><i class="fas fa-thin fa-bars"></i></a></li>-->
                <li class="user"><a href="#"><i class="fas fa-thin fa-user"></i></a>
                 <ul>
-                    <li><a href="/user/login">Ingrese</a></li>
-                    <li><a href="/user/register">Registrese</a></li>
-                    <li><a href="#">Perfil</a></li>
+                    <li><a href="/user/login">Ingresar</a></li>
+                    <li><a href="/user/register">Registrarse</a></li>
+                    <li><a href="/user/profile">Perfil</a></li>
+                    <li><a href="/logout">Cerrar Sesion</a></li>
                     <li><a href="/product/create">Crear producto</a></li>
                     <li><a href="/product/edit">Editar producto</a></li>
                     <li><a href="/product/delete">Eliminar producto</a></li>
