+++ src/views/user/register.ejs
@@ -14,7 +14,7 @@
 
         <div class="login-box">
             <div class="formulario">
-                <form action="/register" method="POST" enctype="multipart/form-data"><h4>Registrate</h4>
+                <form action="/register" method="POST" enctype="multipart/form-data"><h4>Registrarse</h4>
                     <div class="nombre">
 
                         <input type="text" placeholder="Nombre y apellido" name="fullName"> </input>
